requirements.txt
plexapi
requests
tmdbapis
pyyaml
setuptools
schedule
pytz
